Don't shut down sidejob_supervisor upon receiving a spurious EXIT.
When sidejob_supervisor received {'EXIT', Pid, Reason} it checked if
the EXIT signal was sent by one of its children, if so it removed the
child from its state.

If the EXIT came from another Pid sidejob_supervisor terminated, I
guess because of the assumption that in that case Pid was the parent
telling sidejob_supervisor to shut down. Actually this case is already
handled in gen_server itself.

The problem was that if starting a child using apply(Mod, Fun, Args)
failed but a child process was already spawned and linked (which is
what happens in i.e. gen_server:start_link) sidejob_supervisor would
receive an EXIT signal but because the pid wasn't registered as one of
its children it then terminated.

This commit changes behavior to be in accordance with OTP's supervisor
which just ignores EXIT signals from unknown Pids, and delegates
handling the EXIT from its parent to gen_server.