Quviq fix for sidejob_eqc
Also disables supervisor_eqc's running of prop_par. Sidejob has at least
one race condition, but probably many.

For example
https://github.com/basho/sidejob/blob/develop/src/sidejob.erl#L144 has a
fairly small window "check then act" race that allows more than limit
workers.

The larger race is between
https://github.com/basho/sidejob/blob/develop/src/sidejob.erl#L150 and
https://github.com/basho/sidejob/blob/develop/src/sidejob_worker.erl#L201

The first sets the `usage` based on how many want to start a worker, and
the second sets the usage to how many workers actually started. The
actual limit on the number of processes in the worse case is actually
N(N+1)/2 since LIMIT processes could set the ETS counter, then 1 gets
created and set the counter to 1, and then LIMIT-1 update the ETS
counter, and then 1 get created and set the counter to 2, and LIMIT-2
update the counter, etc etc etc.

The test in supervisor_eqc has been fuzzed to reflect this. There is
also a bug in which_children. Counterexamples for all the races and bugs
are added in this commit so that they can be used to validate future
attempts to fix sidejob.